+++ daisy/messages/acquire_block.py
