+++ daisy/messages/message.py
