+++ daisy/tcp/tcp_client.py
@@ -6,7 +6,7 @@ import tornado.tcpclient
 
 from .io_looper import IOLooper
 from .tcp_stream import TCPStream
-from daisy.message import ExceptionMessage
+from daisy.messages import ExceptionMessage
 
 logger = logging.getLogger(__name__)
